# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "윤정 법률사무소 API",
    "version": "1.0.0",
    "description": "윤정 법률사무소 홈페이지 백엔드 REST API. 모든 응답은 { data, error, meta } 형식입니다."
  },
  "servers": [
    { "url": "/api", "description": "기본 API 서버" }
  ],
...
  "paths": {
...
    "/portal/register": {
      "post": {
        "tags": ["Portal"],
        "summary": "포털 회원가입",
        "description": "이메일, 비밀번호(8자 이상), 이름, 전화번호로 가입합니다. 전화번호로 기존 고객과 자동 매칭됩니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name", "phone"],
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "이메일 주소" },
                  "password": { "type": "string", "minLength": 8, "description": "비밀번호 (8자 이상)" },
                  "name": { "type": "string", "description": "이름" },
                  "phone": { "type": "string", "description": "전화번호 (한국 형식)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "email": { "type": "string" },
                            "clientId": { "type": "string", "nullable": true }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "이미 가입된 이메일",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          }
        }
      }
    },
    "/portal/login": {
      "post": {
        "tags": ["Portal"],
        "summary": "포털 로그인",
        "description": "이메일/비밀번호로 인증 후 포털 세션 토큰(x-portal-token)을 반환합니다. 세션 TTL 24시간.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": { "type": "string", "description": "포털 세션 토큰 (x-portal-token 헤더에 사용)" },
                            "user": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string" },
                                "email": { "type": "string" },
                                "clientId": { "type": "string", "nullable": true }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "비활성화된 계정",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          }
        }
      }
    },
...
    "/portal/me": {
      "get": {
        "tags": ["Portal"],
        "summary": "현재 포털 사용자 정보",
        "security": [{ "portalAuth": [] }],
        "responses": {
          "200": {
            "description": "사용자 및 고객 정보",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string" },
                                "email": { "type": "string" },
                                "clientId": { "type": "string", "nullable": true }
                              }
                            },
                            "client": { "$ref": "#/components/schemas/Client" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/portal/cases": {
      "get": {
        "tags": ["Portal"],
        "summary": "내 사건 목록 조회",
        "security": [{ "portalAuth": [] }],
        "responses": {
          "200": {
            "description": "사건 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CaseFile" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/portal/cases/{id}": {
      "get": {
        "tags": ["Portal"],
        "summary": "사건 상세 조회 (문서 + 최근 메시지 포함)",
        "description": "본인 사건만 조회 가능합니다. 관련 문서와 최근 메시지 10건을 포함합니다.",
        "security": [{ "portalAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "responses": {
          "200": {
            "description": "사건 상세",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/CaseFile" },
                            {
                              "type": "object",
                              "properties": {
                                "documents": { "type": "array", "items": { "type": "object" } },
                                "messages": { "type": "array", "items": { "$ref": "#/components/schemas/CaseMessage" } }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/portal/cases/{id}/messages": {
      "get": {
        "tags": ["Portal"],
        "summary": "사건 메시지 목록 (페이지네이션)",
        "security": [{ "portalAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "페이지 번호" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "페이지당 항목 수 (최대 100)" }
        ],
        "responses": {
          "200": {
            "description": "메시지 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CaseMessage" }
                        },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "tags": ["Portal"],
        "summary": "의뢰인 메시지 전송",
        "security": [{ "portalAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "description": "메시지 내용" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "전송된 메시지",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseMessage" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/portal/admin/cases": {
      "get": {
        "tags": ["Portal Admin"],
        "summary": "전체 사건 목록 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "페이지 번호" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "페이지당 항목 수 (최대 100)" }
        ],
        "responses": {
          "200": {
            "description": "사건 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CaseFile" }
                        },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Portal Admin"],
        "summary": "사건 생성 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clientId", "title"],
                "properties": {
                  "clientId": { "type": "string", "format": "uuid", "description": "고객 ID" },
                  "title": { "type": "string", "description": "사건 제목" },
                  "lawyerId": { "type": "string", "nullable": true, "description": "담당 변호사 ID" },
                  "description": { "type": "string", "nullable": true, "description": "사건 설명" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 사건",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseFile" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/portal/admin/cases/{id}": {
      "patch": {
        "tags": ["Portal Admin"],
        "summary": "사건 수정 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "description": "사건 상태" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "lawyerId": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 사건",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseFile" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/portal/admin/cases/{id}/messages": {
      "post": {
        "tags": ["Portal Admin"],
        "summary": "변호사 메시지 전송 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "description": "메시지 내용" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "전송된 메시지",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseMessage" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "관리자 인증 토큰 (POST /admin-users/login에서 발급)"
      },
      "portalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-portal-token",
        "description": "포털 세션 토큰 (POST /portal/login에서 발급, TTL 24시간)"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "data": { "description": "응답 데이터 (성공 시)", "nullable": true },
          "error": { "type": "string", "description": "에러 메시지 (실패 시)", "nullable": true },
          "meta": { "type": "object", "description": "부가 정보 (페이지네이션 등)", "nullable": true }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "description": "현재 페이지" },
          "limit": { "type": "integer", "description": "페이지당 항목 수" },
          "total": { "type": "integer", "description": "전체 항목 수" },
          "totalPages": { "type": "integer", "description": "전체 페이지 수" }
        }
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string", "nullable": true },
          "source": { "type": "string", "nullable": true, "description": "유입 경로" },
          "memo": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CaseFile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "status": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "lawyerId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CaseMessage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "caseFileId": { "type": "string", "format": "uuid" },
          "senderId": { "type": "string", "nullable": true },
          "senderType": { "type": "string", "enum": ["client", "lawyer"], "description": "발신자 유형" },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "필수 항목이 누락되었습니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "인증이 필요합니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "리소스를 찾을 수 없습니다" }
                  }
                }
              ]
            }
          }
        }
      }
    }
  }
}